update test cases
