+++ examples/base.py
@@ -1,0 +1,14 @@
+import struct
+
+def parse_bip32_path(path):
+	if len(path) == 0:
+		return ""
+	result = ""
+	elements = path.split('/')
+	for pathElement in elements:
+		element = pathElement.split('\'')
+		if len(element) == 1:
+			result = result + struct.pack(">I", int(element[0])).hex()
+		else:
+			result = result + struct.pack(">I", 0x80000000 | int(element[0])).hex()
+	return result
