+++ examples/getPublicKey.py
@@ -2,20 +2,7 @@
 
 from ledgerblue.comm import getDongle
 import argparse
-import struct
-
-def parse_bip32_path(path):
-	if len(path) == 0:
-		return ""
-	result = ""
-	elements = path.split('/')
-	for pathElement in elements:
-		element = pathElement.split('\'')
-		if len(element) == 1:
-			result = result + struct.pack(">I", int(element[0])).hex()
-		else:
-			result = result + struct.pack(">I", 0x80000000 | int(element[0])).hex()
-	return result
+from base import parse_bip32_path
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--path', help="BIP32 path to retrieve. e.g. \"44'/195'/0'/0/0\".")
